\name{PJsTools-package}
\alias{PJsTools-package}
\alias{PJsTools}
\docType{package}
\title{
What the package does (short line)
PJ's Tools
}
\description{
Some of my favorite and custom functions
}
\details{
\tabular{ll}{
Package: \tab PJsTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-15\cr
License: \tab GPL\cr
}
TODO
}
\author{
Pieter-Jan Volders <pieterjan.volders@ugent.be>
}
